# Convert ReRegisterForFinalize to QCALL