Contributes to #95695

Benchmark with both finalizable and non-finalizable objects:

```csharp
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkRunner.Run<Perf_ReRegisterForFinalize>(args: args);

public class Perf_ReRegisterForFinalize
{
    private static readonly object notFinializableObject = new();
    private static readonly object finializableObject = new FinializableObject();

    [Benchmark]
    public void NotFinializable() => GC.ReRegisterForFinalize(notFinializableObject);

    [Benchmark]
    public void Finializable() => GC.ReRegisterForFinalize(finializableObject);
}

class FinializableObject
{
    ~FinializableObject() {}
}
```

```txt
BenchmarkDotNet v0.13.12, macOS Sonoma 14.5 (23F79) [Darwin 23.5.0]
Apple M2, 1 CPU, 8 logical and 8 physical cores
.NET SDK 8.0.300
  [Host]     : .NET 8.0.5 (8.0.524.21615), Arm64 RyuJIT AdvSIMD
  Job-UXNDHB : .NET 9.0.0 (42.42.42.42424), Arm64 RyuJIT AdvSIMD

Toolchain=CoreRun  
```

Finializable

| Version          | Mean      | Error     | StdDev    |
|---------------- |----------:|----------:|----------:|
| [main](https://github.com/dotnet/runtime/commit/39ecbe0ed850fa6d6d4f0ebe1917bca1e154e82d)    | 15.057 ns | 0.1719 ns | 0.2466 ns |
| PR    | 12.4418 ns | 0.1961 ns | 0.2812 ns |

NotFinializable

| Version          | Mean      | Error     | StdDev    |
|---------------- |----------:|----------:|----------:|
| [main](https://github.com/dotnet/runtime/commit/39ecbe0ed850fa6d6d4f0ebe1917bca1e154e82d) |  1.315 ns | 0.0019 ns | 0.0017 ns |
| PR |  0.3360 ns | 0.0024 ns | 0.0021 ns |